- In the TEXTUREx editor, any part of a texture outside the texture bounds is now drawn semi-transparent
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@437 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/GLTexture.cpp
@@ -277,7 +277,7 @@ bool GLTexture::bind() {
 	return true;
 }
 
-bool GLTexture::draw2d(int32_t x, int32_t y) {
+bool GLTexture::draw2d(double x, double y) {
 	// Can't draw if texture not loaded
 	if (!loaded)
 		return false;
